+++ src/main/java/glue/test/Test.java
@@ -1,5 +1,0 @@
-package glue.test;
-
-class Test {
-
-}