+++ CHANGES.md
@@ -7,6 +7,8 @@ Release date: Not released
  - Glue logger module implementation with the default CDI provider for SLF4J instances, using Logback as the provider 
  - Configuration API implementation, allowing transparent and decoupled access to configuration values by Glue modules
  - Configuration resolver module implementation using Apache Deltaspike as the provider
+ - Implementation of the initial version of the Web Container API
+ - Initial version of the Web Container using Jetty as the provider for the implementation
  
 ### Minor changes and enhancements